Support for image data file production (#191)
Updates the Image Builder tool to define commands that aid in the production of a file that contains info about images produced during a build. The end goal of all this is to be able to automatically queue a build whenever a base image changes and have that build only target the images that reference that changed base image.