Enable manifest filtering for RebuildStaleImages command (#195)
The RebuildStaleImages command currently enumerates all images in the manifest and attempts to pull them to get their base image digest value. This can't be done when there is a mix of Linux and Windows images because you can't pull a Windows image from a Linux Docker host, for example. For that reason, there needs to be filtering options for the command that are applied to the manifests that are consumed.