Image info file support for CopyAcrImages cmd (#239)
Updates the CopyAcrImages command to optionally incorporate the data from the image info file. This is necessary to handle image tags that make use of dynamically generated values, such as timestamps. Without these changes, Image Builder would calculate a new tag value each time it is invoked. This is obviously not good when we're trying to reference the tag of an image that was built in a separate build step.

By making use of the image info file that was generated when the image was built, we know the exact static value of the tag that needs to be referenced. So when the CopyAcrImages command runs and finds a Platform to process, it looks up the tag value in the image info file instead of using the value that was calculated when parsing the manifest.